# src/main/resources/swagger.yaml
openapi: 3.0.3
info:
  title: 5Hertz
  version: 1.0.4
  description: >
    Documentacion de la API productiva del sistema de produccion desarrollado por 5Hertz orientado a PYMES alimenticias.
servers:
  - url: http://localhost:8080
    description: Servidor local de desarrollo
...
paths:
...
  /stock/habilitar-producto:
    put:
      summary: Habilitar producto en almacen
      tags: [StockAlmacen]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HabilitarProductoRequest'
      responses:
        '200':
          description: Producto habilitado correctamente.
  /stock/incrementar:
    put:
      summary: Incrementar stock total
      tags: [StockAlmacen]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModificarStockRequest'
      responses:
        '200':
          description: Stock incrementado correctamente.
  /stock/reducir:
    put:
      summary: Reducir stock total
      tags: [StockAlmacen]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModificarStockRequest'
      responses:
        '200':
          description: Stock reducido correctamente.
...
  /stock/consultar-producto-almacen:
    get:
      summary: Consultar stock de un producto en un almacen
      tags: [StockAlmacen]
      parameters:
        - name: sku
          in: query
          required: true
          schema:
            type: string
        - name: idAlmacen
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Stock obtenido correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockAlmacen'
        '404':
          description: Producto o almacen no encontrado.
...
components:
  schemas:
...
    HabilitarProductoRequest:
      type: object
      properties:
        idAlmacen:
          type: integer
        sku:
          type: string
    ModificarStockRequest:
      type: object
      properties:
        idAlmacen:
          type: integer
        sku:
          type: string
        cantidad:
          type: number
    StockAlmacen:
      type: object
      properties:
        idStock:
          type: integer
        idAlmacen:
          type: integer
        sku:
          type: string
        cantidadMinima:
          type: number
        stockReservado:
          type: number
        stockDisponible:
          type: number
        stockTotal:
          type: number
...